+++ temba/channels/models.py
@@ -140,6 +140,12 @@ class ChannelType(metaclass=ABCMeta):
             return UpdateChannelForm
         return self.update_form
 
+    def check_credentials(self, config: dict) -> bool:
+        """
+        Called to check the credentials passed are valid
+        """
+        return True
+
     def activate(self, channel):
         """
         Called when a channel of this type has been created. Can be used to setup things like callbacks required by the
@@ -340,7 +346,7 @@ class Channel(LegacyUUIDMixin, TembaModel, DependencyMixin):
         help_text=_("We will send email alerts to this address if experiencing issues sending"),
     )
 
-    config = JSONAsTextField(null=True, default=dict)
+    config = JSONAsTextField(default=dict)
     schemes = ArrayField(models.CharField(max_length=16), default=_get_default_channel_scheme)
     role = models.CharField(max_length=4, default=DEFAULT_ROLE)
     parent = models.ForeignKey("self", on_delete=models.PROTECT, null=True)
@@ -713,6 +719,9 @@ class Channel(LegacyUUIDMixin, TembaModel, DependencyMixin):
     def is_new(self):
         # is this channel newer than an hour
         return self.created_on > timezone.now() - timedelta(hours=1) or not self.last_sync
+
+    def check_credentials(self) -> bool:
+        return self.type.check_credentials(self.config)
 
     def release(self, user, *, trigger_sync: bool = True):
         """