+++ temba/msgs/models.py
@@ -161,6 +161,9 @@ class Media(models.Model):
 
         process_upload(self)
 
+    def __str__(self) -> str:
+        return f"{self.content_type}:{self.url}"
+
     class Meta:
         indexes = [
             models.Index(name="media_originals_by_org", fields=["org", "-created_on"], condition=Q(original=None))
@@ -243,7 +246,9 @@ class Broadcast(models.Model):
             for lang, atts in attachments.items():
                 if lang not in translations:
                     translations[lang] = {}
-                translations[lang]["attachments"] = atts
+
+                # TODO update broadcast sending to allow media objects to stay as UUIDs for longer
+                translations[lang]["attachments"] = [str(m) for m in atts]
 
         broadcast = cls.objects.create(
             org=org,
@@ -464,13 +469,11 @@ class Msg(models.Model):
     DIRECTION_OUT = "O"
     DIRECTION_CHOICES = ((DIRECTION_IN, "Incoming"), (DIRECTION_OUT, "Outgoing"))
 
-    TYPE_INBOX = "I"  # to be replaced with T
-    TYPE_FLOW = "F"  # to be replaced with T
+    TYPE_INBOX = "I"  # no longer used
+    TYPE_FLOW = "F"  # no longer used
     TYPE_TEXT = "T"
     TYPE_VOICE = "V"
     TYPE_CHOICES = (
-        (TYPE_INBOX, "Inbox Message"),
-        (TYPE_FLOW, "Flow Message"),
         (TYPE_TEXT, "Text Message"),
         (TYPE_VOICE, "Voice Message"),
     )
@@ -502,7 +505,7 @@ class Msg(models.Model):
     MAX_ATTACHMENT_LEN = settings.MSG_ATTACHMENT_SIZE  # max attachments allowed in a message
 
     id = models.BigAutoField(primary_key=True)
-    uuid = models.UUIDField(null=True, default=uuid4)
+    uuid = models.UUIDField(default=uuid4)
     org = models.ForeignKey(Org, on_delete=models.PROTECT, related_name="msgs", db_index=False)
 
     # message destination
@@ -526,7 +529,7 @@ class Msg(models.Model):
     sent_on = models.DateTimeField(null=True)
     queued_on = models.DateTimeField(null=True)
 
-    msg_type = models.CharField(max_length=1, choices=TYPE_CHOICES, null=True)
+    msg_type = models.CharField(max_length=1, choices=TYPE_CHOICES)
     direction = models.CharField(max_length=1, choices=DIRECTION_CHOICES)
     status = models.CharField(max_length=1, choices=STATUS_CHOICES, default=STATUS_PENDING, db_index=True)
     visibility = models.CharField(max_length=1, choices=VISIBILITY_CHOICES, default=VISIBILITY_VISIBLE)